python/requirements.txt
pymupdf
pix2tex
Pillow
openai
cloudinary
google-cloud-vision
psycopg2-binary
python-dotenv
regex
